requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
diff-match-patch==20230430
Django==4.2.13
django-import-export==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
tablib==3.5.0
typing-extensions==4.12.2
